fix(ci): repair failing quality gates
