+++ native/rust/crates/ripdpi-monitor-engine/src/session/lifecycle.rs
@@ -1,22 +1,17 @@
-use std::sync::atomic::{AtomicBool, Ordering};
+mod reporting;
+mod start;
+
+use std::sync::atomic::AtomicBool;
 use std::sync::{Arc, Mutex, MutexGuard};
 use std::thread::JoinHandle;
-use std::time::{Duration, Instant};
+use std::time::Instant;
 
 use rustls::client::danger::ServerCertVerifier;
 
 use crate::execution::UnavailableCandidateRuntimeLauncher;
 use crate::platform::NoopMonitorPlatformBridge;
-use crate::types::{
-    EngineScanRequestWire, ScanCompletionKind, ScanReportDisposition, ScanRequest, ScanTerminationReason, SharedState,
-};
+use crate::types::{ScanTerminationReason, SharedState};
 use crate::{CandidateRuntimeLauncher, MonitorPlatformBridge};
-
-use super::log_level::parse_native_log_level;
-use super::reaper::WORKER_REAPER;
-use super::validation::ValidatedScanRequest;
-use super::wire_json::{passive_events_to_json, progress_to_json, report_to_json};
-use super::worker::{ScanWorkerConfig, join_finished_worker_locked, spawn_scan_worker};
 
 #[derive(Default)]
 struct ScanControl {
@@ -25,44 +20,6 @@ struct ScanControl {
     terminal_report_delivered: bool,
     start_in_progress: bool,
 }
-
-enum StartMarkerState<'a> {
-    NotAdmitted,
-    Admitted(&'a Mutex<ScanControl>),
-}
-
-struct StartInProgressGuard<'a> {
-    state: StartMarkerState<'a>,
-}
-
-impl<'a> StartInProgressGuard<'a> {
-    fn new() -> Self {
-        Self { state: StartMarkerState::NotAdmitted }
-    }
-
-    fn mark_admitted(&mut self, scan_control: &'a Mutex<ScanControl>) {
-        self.state = StartMarkerState::Admitted(scan_control);
-    }
-
-    fn clear(&mut self) {
-        let StartMarkerState::Admitted(scan_control) =
-            std::mem::replace(&mut self.state, StartMarkerState::NotAdmitted)
-        else {
-            return;
-        };
-        let mut control = scan_control.lock().unwrap_or_else(std::sync::PoisonError::into_inner);
-        control.start_in_progress = false;
-    }
-}
-
-impl Drop for StartInProgressGuard<'_> {
-    fn drop(&mut self) {
-        self.clear();
-    }
-}
-
-type ScanWorkerSpawner =
-    fn(Arc<Mutex<SharedState>>, Arc<AtomicBool>, String, ScanRequest, ScanWorkerConfig) -> JoinHandle<()>;
 
 pub struct MonitorSession {
     // Lock order: worker -> active_session_id -> scan_control -> cancellation_reason -> shared.
@@ -123,166 +80,6 @@ impl MonitorSession {
         }
     }
 
-    pub fn start_scan(&self, session_id: String, request: EngineScanRequestWire) -> Result<(), String> {
-        self.start_scan_with_spawner(session_id, request, spawn_scan_worker)
-    }
-
-    fn start_scan_with_spawner(
-        &self,
-        session_id: String,
-        request: EngineScanRequestWire,
-        spawn_worker: ScanWorkerSpawner,
-    ) -> Result<(), String> {
-        let request = ValidatedScanRequest::try_from(request)?;
-        let native_log_level = parse_native_log_level(request.as_wire().native_log_level.as_deref())?;
-        {
-            let mut worker_guard = self.worker.lock().map_err(|_| "monitor worker poisoned".to_string())?;
-            join_finished_worker_locked(&mut worker_guard);
-            if worker_guard.is_some() {
-                return Err("diagnostics scan already running".to_string());
-            }
-        }
-        let scan_deadline =
-            Instant::now() + Duration::from_millis(request.as_wire().scan_deadline_ms.unwrap_or(360_000));
-        let log_context = request.as_wire().log_context.clone();
-        let mut start_guard = StartInProgressGuard::new();
-        {
-            let mut active_session_id =
-                self.active_session_id.lock().map_err(|_| "monitor session id poisoned".to_string())?;
-            let mut scan_control = self.scan_control.lock().map_err(|_| "monitor scan control poisoned".to_string())?;
-            if scan_control.start_in_progress {
-                return Err("diagnostics scan already running".to_string());
-            }
-            *active_session_id = Some(session_id.clone());
-            *scan_control = ScanControl {
-                deadline: Some(scan_deadline),
-                checkpoint_report_delivered: false,
-                terminal_report_delivered: false,
-                start_in_progress: true,
-            };
-            start_guard.mark_admitted(&self.scan_control);
-            // The scan-control lock is the admission barrier. A cancellation
-            // racing after admission blocks on this lock and publishes its
-            // flag/cause only after the prior generation has been reset.
-            self.cancel.store(false, Ordering::Release);
-            *self.cancellation_reason.lock().map_err(|_| "monitor cancellation state poisoned".to_string())? = None;
-            let mut shared = self.lock_shared_state_recovering();
-            shared.progress = None;
-            shared.report = None;
-            shared.checkpoint_report = None;
-            shared.log_context = log_context;
-            shared.terminal_session_id = None;
-        }
-        self.platform_bridge.clear_passive_events(&session_id);
-        let domain_request = request.into();
-        let worker_config = ScanWorkerConfig::new(
-            scan_deadline,
-            self.cancellation_reason.clone(),
-            self.tls_verifier.clone(),
-            self.platform_bridge.clone(),
-            self.candidate_runtime_launcher.clone(),
-            native_log_level,
-        );
-        let mut worker_guard = self.worker.lock().map_err(|_| "monitor worker poisoned".to_string())?;
-        join_finished_worker_locked(&mut worker_guard);
-        if worker_guard.is_some() {
-            return Err("diagnostics scan already running".to_string());
-        }
-        let worker = std::panic::catch_unwind(std::panic::AssertUnwindSafe(|| {
-            spawn_worker(self.shared.clone(), self.cancel.clone(), session_id, domain_request, worker_config)
-        }))
-        .map_err(|_| "diagnostics worker spawn panicked".to_string())?;
-        *worker_guard = Some(worker);
-        drop(start_guard);
-        Ok(())
-    }
-
-    pub fn cancel_scan(&self) {
-        let reason = self.scan_control.lock().ok().map_or(ScanTerminationReason::UserCancelled, |scan_control| {
-            if scan_control.deadline.is_some_and(|deadline| Instant::now() >= deadline) {
-                ScanTerminationReason::DeadlineExceeded
-            } else {
-                ScanTerminationReason::UserCancelled
-            }
-        });
-        if let Ok(mut cancellation_reason) = self.cancellation_reason.lock()
-            && cancellation_reason.is_none()
-        {
-            *cancellation_reason = Some(reason);
-        }
-        self.cancel.store(true, Ordering::Release);
-    }
-
-    pub fn poll_progress_json(&self) -> Result<Option<String>, String> {
-        let shared = self.lock_shared_state_recovering();
-        progress_to_json(shared.progress.as_ref())
-    }
-
-    pub fn take_report_json(&self) -> Result<Option<String>, String> {
-        self.try_join_worker();
-        let mut scan_control = self.scan_control.lock().map_err(|_| "monitor scan control poisoned".to_string())?;
-        if scan_control.terminal_report_delivered {
-            return Ok(None);
-        }
-        // Ordering: observes cancellation published by cancel_scan before exposing a checkpoint.
-        let cancellation_requested = self.cancel.load(Ordering::Acquire);
-        let cancellation_reason =
-            self.cancellation_reason.lock().map_err(|_| "monitor cancellation state poisoned".to_string())?.clone();
-        let mut shared = self.lock_shared_state_recovering();
-        let scan_finished = shared.progress.as_ref().is_none_or(|progress| progress.is_finished);
-        if !scan_finished && !cancellation_requested {
-            return Ok(None);
-        }
-        if cancellation_requested && !scan_finished {
-            if scan_control.checkpoint_report_delivered {
-                return Ok(None);
-            }
-            let mut checkpoint = shared.checkpoint_report.take();
-            if let Some(report) = checkpoint.as_mut() {
-                report.completion_kind = ScanCompletionKind::PartialResults;
-                report.report_disposition = ScanReportDisposition::Checkpoint;
-                report.termination_reason = cancellation_reason;
-                report.finished_at = crate::util::now_ms();
-            }
-            let json = report_to_json(checkpoint.as_ref())?;
-            scan_control.checkpoint_report_delivered = json.is_some();
-            return Ok(json);
-        }
-        let report = shared.report.take();
-        let json = report_to_json(report.as_ref())?;
-        scan_control.terminal_report_delivered = json.is_some();
-        Ok(json)
-    }
-
-    pub fn poll_passive_events_json(&self) -> Result<Option<String>, String> {
-        let session_id = self.active_session_id.lock().map_err(|_| "monitor session id poisoned".to_string())?.clone();
-        let events = session_id.as_deref().map(|id| self.platform_bridge.drain_passive_events(id)).unwrap_or_default();
-        passive_events_to_json(events)
-    }
-
-    /// Cancel the active scan and retire its worker without blocking the caller.
-    ///
-    /// An unfinished worker is joined by the process-wide diagnostics reaper.
-    /// This keeps JNI teardown bounded even when a probe is inside blocking I/O;
-    /// the worker still owns its state until it exits and is reaped.
-    pub fn destroy(&self) {
-        self.cancel_scan();
-        let handle = self.worker.lock().ok().and_then(|mut worker_guard| worker_guard.take());
-        if let Some(handle) = handle
-            && let Err(handle) = WORKER_REAPER.reap(handle)
-        {
-            log::error!("detaching diagnostics worker because the bounded reaper is saturated or unavailable");
-            drop(handle);
-        }
-    }
-
-    fn try_join_worker(&self) {
-        let Ok(mut worker_guard) = self.worker.lock() else {
-            return;
-        };
-        join_finished_worker_locked(&mut worker_guard);
-    }
-
     fn lock_shared_state_recovering(&self) -> MutexGuard<'_, SharedState> {
         // SharedState is an owned diagnostics snapshot store. A writer panic can
         // poison the mutex even after panic recovery published a safe terminal
@@ -299,10 +96,12 @@ impl MonitorSession {
 mod tests {
     use super::*;
     use crate::types::{
-        DIAGNOSTICS_ENGINE_SCHEMA_VERSION, DiagnosticProfileFamily, NativeSessionEvent, ProbeResult,
-        ScanCompletionKind, ScanKind, ScanPathMode, ScanProgress, ScanReport, ScanTerminationReason,
+        DIAGNOSTICS_ENGINE_SCHEMA_VERSION, DiagnosticProfileFamily, EngineScanRequestWire, NativeSessionEvent,
+        ProbeResult, ScanCompletionKind, ScanKind, ScanPathMode, ScanProgress, ScanReport, ScanReportDisposition,
+        ScanTerminationReason,
     };
     use ripdpi_telemetry::recorder::RecorderSnapshot;
+    use std::sync::atomic::Ordering;
     use std::sync::{Barrier, Weak, mpsc};
     use std::thread;
     use std::time::{Duration, Instant};
