+++ native/rust/crates/ripdpi-monitor-engine/src/session/worker.rs
@@ -1,3 +1,5 @@
+mod execution;
+
 use std::sync::atomic::AtomicBool;
 use std::sync::{Arc, Mutex};
 use std::thread::{self, JoinHandle};
@@ -6,13 +8,12 @@ use std::time::Instant;
 use log::LevelFilter;
 use rustls::client::danger::ServerCertVerifier;
 
-use crate::engine::run_engine_scan;
 use crate::types::{ScanRequest, ScanTerminationReason, SharedState};
 use crate::{CandidateRuntimeLauncher, MonitorPlatformBridge};
 
 use super::panic_state::record_panic_terminal_state;
 
-pub(super) struct ScanWorkerConfig {
+pub(crate) struct ScanWorkerConfig {
     scan_deadline: Instant,
     cancellation_reason: Arc<Mutex<Option<ScanTerminationReason>>>,
     tls_verifier: Option<Arc<dyn ServerCertVerifier>>,
@@ -22,7 +23,7 @@ pub(super) struct ScanWorkerConfig {
 }
 
 impl ScanWorkerConfig {
-    pub(super) fn new(
+    pub(crate) fn new(
         scan_deadline: Instant,
         cancellation_reason: Arc<Mutex<Option<ScanTerminationReason>>>,
         tls_verifier: Option<Arc<dyn ServerCertVerifier>>,
@@ -54,7 +55,7 @@ pub(super) fn spawn_scan_worker(
     thread::spawn(move || {
         let started_at = crate::util::now_ms();
         let result = std::panic::catch_unwind(std::panic::AssertUnwindSafe(|| {
-            run_scan(shared, cancel, session_id, request, config);
+            execution::run_scan(shared, cancel, session_id, request, config);
         }));
         if let Err(panic_payload) = result {
             record_panic_terminal_state(shared_panic, session_id_panic, request_panic, started_at, panic_payload);
@@ -73,28 +74,6 @@ pub(super) fn join_finished_worker_locked(worker_guard: &mut Option<JoinHandle<(
             let _ = handle.join();
         }
     }
-}
-
-fn run_scan(
-    shared: Arc<Mutex<SharedState>>,
-    cancel: Arc<AtomicBool>,
-    session_id: String,
-    request: ScanRequest,
-    config: ScanWorkerConfig,
-) {
-    let _log_scope = config
-        .native_log_level
-        .map(|level| config.platform_bridge.scoped_log_level("diagnostics_native".to_string(), level));
-    run_engine_scan(
-        shared,
-        cancel,
-        session_id,
-        request,
-        config.scan_deadline,
-        config.cancellation_reason,
-        config.tls_verifier,
-        config.candidate_runtime_launcher,
-    );
 }
 
 #[cfg(test)]
